requirements.txt
pymongo
asyncio
datetime
loguru
aiohttp
pandas
python-dotenv
fastapi
requests
pydantic
uvicorn
yfinance
backtesting